Move files to nested module directory to avoid conflict with top-level config.rb file in other gems
