+++ lib/proj.rb
@@ -1,22 +1,22 @@
 # encoding: UTF-8
 
 require_relative 'api/api'
-require_relative './config'
+require_relative 'proj/config'
 
-require_relative './area'
-require_relative './context'
-require_relative './coordinate'
-require_relative './ellipsoid'
-require_relative './error'
-require_relative './point'
-require_relative './prime_meridian'
-require_relative './unit'
+require_relative 'proj/area'
+require_relative 'proj/context'
+require_relative 'proj/coordinate'
+require_relative 'proj/ellipsoid'
+require_relative 'proj/error'
+require_relative 'proj/point'
+require_relative 'proj/prime_meridian'
+require_relative 'proj/unit'
 
-require_relative './pj_object'
-require_relative './operation'
-require_relative './crs'
-require_relative './projection'
-require_relative './transformation'
+require_relative 'proj/pj_object'
+require_relative 'proj/operation'
+require_relative 'proj/crs'
+require_relative 'proj/projection'
+require_relative 'proj/transformation'
 
 module Proj
   def self.info
