+++ lib/proj/area.rb
