+++ lib/proj/config.rb
