+++ lib/proj/coordinate.rb
