+++ lib/proj/ellipsoid.rb
