+++ lib/proj/error.rb
