+++ lib/proj/pj_object.rb
