+++ lib/proj/point.rb
